Refactor backup to submodules
