+++ core/src/encryption/mod.rs
@@ -13,7 +13,7 @@ mod key_name;
 mod keychains;
 mod root_backup_key;
 
-pub use crate::encryption::{
+pub(crate) use crate::encryption::{
     backup_password::BackupPassword,
     encryption_key::{DataEncryptionKey, KeyEncryptionKey},
     encryption_output::EncryptionOutput,
